requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
djangorestframework==3.14.0
drf-yasg==1.21.8
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
Markdown==3.5.2
MarkupSafe==3.0.2
mysqlclient==2.2.6
openapi-codec==1.3.2
packaging==24.2
pillow==10.2.0
psycopg2-binary==2.9.9
pyclean==3.0.0
pytz==2024.1
PyYAML==6.0.2
requests==2.31.0
simplejson==3.19.3
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
